# Optimize binary match from 10% up to 70x