--- a/lib/stdlib/test/stdlib.spec
+++ b/lib/stdlib/test/stdlib.spec
@@ -1,4 +1,4 @@
 {suites,"../stdlib_test",all}.
 {skip_groups,"../stdlib_test",stdlib_bench_SUITE,
-             [base64,gen_server,gen_statem,unicode],
+             [binary,base64,gen_server,gen_statem,unicode],
              "Benchmark only"}.