@josevalim 2018-04-29
# Benchmarks for BOYER_MOORE_REDUCTION_FACTOR=10 and MEMCHR_REDUCTION_FACTOR=10

## Large binary with no match

subject = "ugbcfuysabfuqyfikgfsdalpaskfhgjsdgfjwsalp" x 1_000_000
pattern = "o"

### With this patch

binary:match/2 not precompiled: 206962.24us
binary:match/2 precompiled: 206611.66us

binary:split/2 not precompiled: 206123.68us
binary:split/2 precompiled: 206414.64us

binary:matches/2 not precompiled: 212258.74us
binary:matches/2 precompiled: 213588.46us

binary:split/2 not precompiled: 213542.1us
binary:split/2 precompiled: 212566.94us

### Without this patch

binary:match/2 not precompiled: 223504.4us
binary:match/2 precompiled: 222565.64us

binary:split/2 not precompiled: 222964.44us
binary:split/2 precompiled: 224047.88us

binary:matches/2 not precompiled: 224721.34us
binary:matches/2 precompiled: 224562.2us

binary:split/2 not precompiled: 224625.42us
binary:split/2 precompiled: 224629.42us

### Results

About 10% faster.

## Large binary with multiple single byte matches

subject = "abchfkxhfjksdhfjksahjfhasdjkfhdksjfhjkadshfjkdshfjkerpoqkfl\n" x 1_000_000
pattern = "\n"

The goal is to test a single byte happening from time to time.

### With this patch

binary:match/2 not precompiled: 0.96us
binary:match/2 precompiled: 0.44us

binary:split/2 not precompiled: 1.56us
binary:split/2 precompiled: 0.48us

binary:matches/2 not precompiled: 104144.64us
binary:matches/2 precompiled: 99908.9us

binary:split/2 not precompiled: 103840.48us
binary:split/2 precompiled: 100241.76us

### Without this patch

binary:match/2 not precompiled: 3.36us
binary:match/2 precompiled: 1.08us

binary:split/2 not precompiled: 1.06us
binary:split/2 precompiled: 0.66us

binary:matches/2 not precompiled: 445137.84us
binary:matches/2 precompiled: 432233.44us

binary:split/2 not precompiled: 437167.26us
binary:split/2 precompiled: 436216.9us

### Results

About 4 times faster.

## Large binary with sequential three byte matches

subject = "abc" x 1_000_000
pattern = "abc"

The goal is to test a pattern happening over and over in sequence.

### With this patch

binary:match/2 not precompiled: 1.48us
binary:match/2 precompiled: 0.22us

binary:split/2 not precompiled: 0.8us
binary:split/2 precompiled: 0.32us

binary:matches/2 not precompiled: 109711.36us
binary:matches/2 precompiled: 107959.44us

binary:split/2 not precompiled: 104394.38us
binary:split/2 precompiled: 101978.8us

### Without this patch

binary:match/2 not precompiled: 1.24us
binary:match/2 precompiled: 0.44us

binary:split/2 not precompiled: 0.8us
binary:split/2 precompiled: 0.34us

binary:matches/2 not precompiled: 101125.82us
binary:matches/2 precompiled: 104963.38us

binary:split/2 not precompiled: 101220.5us
binary:split/2 precompiled: 101323.14us

### Results

No measurable difference.