@josevalim 2018-04-29
The previous implementation had a bug where the number of reductions went below zero, never giving the scheduler a chance to stop which affected. I have fixed this bug and updated the benchmark results that now show gains up to 70x.