@josevalim 2018-04-29
Finally, I have pushed a separate commit that adds a reduction factor when using memchr. This speeds up performance about 10% in the no match case. The reasoning is that using memchr is cheaper compared to boyer-moore. I can squash commits but for now they are separate for discussion purposes.