@garazdawi 2018-05-02
Looks good after a cursory look. I'll take a closer look once rc1 is done.

Would it be possible to add the benchmarks that you have written to stdlib_bench_SUITE? so that we can track them and make sure that we don't get any performance regressions.