@garazdawi 2018-05-16
Testcases fail in the stdlib:string_SUITE after applying this pr.