@garazdawi 2018-05-16
Also, we were talking a bit about this optimization. If you want to optimize this even further, you could use memmem on platforms that support it.