@josevalim 2018-05-16
@garazdawi I will look at the tests soon and figure out why they didn't fail locally.

Regarding memmem, I have checked its implementation and the implementation I saw used memchr underneath, in an optimization very similar to the one in this PR. The only difference is that they add a factor: if the pattern you are looking for has more than 32 bytes, they claim the use of BM is better than using memchr. However, I could not reproduce this behaviour, even for larger bytes.

So at the end I thought adding a third path wouldn't be that beneficial.

But talking about optimizations, @michalmuskala and I want to look into optimizing the multi-pattern algorithm in binary:match to use the PCMPESTRI and PCMPESTRM SIMD instructions. Those instructions allow looking up multiple bytes in a string. So it should work similar to the current BM optimization, except we are applying it to a list of patterns.