@garazdawi 2018-05-16
It is the find/split/replace testcases that fail. They seem to fail on all of our test machines, including my own machine. There is also a bunch of other testcases that fail in various applications, the compiler also seems to missbehave on some platforms as it produces invalid code.