@fenollp 2018-05-16
Not exactly on topic:
How realistic could it be to implement SIMD operations for binary comprehensions?
Surely there’re a couple encoder/decoder implementations on the BEAM that would benefit from vectorized `- + * & | ^ << >>` operations? BC semantics seem perfect for such an acceleration.