@garazdawi 2018-05-16
One of the problems with using SIMD in the interpreter is that for it to work we have to sacrifice local tracing on those functions, or make the compiler emit both optimized and traceable code that can be switched during run-time. Regardless, it is quite an undertaking. HiPE on the other hand could (and maybe does? I'm not sure) use SIMD instructions, as it has already sacrificed tracing.