@garazdawi 2018-05-16
Found these backtraces in our debug builds:

```
Thread 1 (Thread 11869):
#0  0x00007f09126e21b5 in raise () from /lib/libc.so.6
#1  0x00007f09126e4fc0 in abort () from /lib/libc.so.6
#2  0x000000000068db84 in erl_assert_error (expr=0x7f8e65 "mem_read > 0", func=0x7f9520 "bm_find_all_non_overlapping", file=0x7f8e40 "beam/erl_bif_binary.c", line=908) at sys/unix/sys.c:940
#3  0x0000000000666bf0 in bm_find_all_non_overlapping (ctx=0x7f08d00bc9a0, haystack=0x7f08cf903ea8 "=PROGRESS REPORT==== 16-May-2018::01:34:29.849613 ===\n    supervisor: {local,sasl_safe_sup}\n    started: [{pid,<0.14176.0>},\n", ' ' , "{id,alarm_handler},\n", ' ' , "{mfargs,{alarm_handler,star"...) at beam/erl_bif_binary.c:908
#4  0x0000000000668704 in do_binary_find (p=0x7f08d19444c8, subject=139675683751098, ctxp=0x7f08d00bc998, pat_bin=0x7f08d1dc2c50, ctx_bin=0x0, res_term=0x7f08d00bc988) at beam/erl_bif_binary.c:1414
#5  0x0000000000669486 in binary_split (p=0x7f08d19444c8, arg1=139675683751098, arg2=139675868277050, arg3=139675868277073) at beam/erl_bif_binary.c:1578
#6  0x0000000000669662 in binary_split_3 (A__p=0x7f08d19444c8, BIF__ARGS=0x7f08d1fc0100, A__I=0x7f08cdd1a7c8) at beam/erl_bif_binary.c:1602
#7  0x000000000044127a in process_main (x_reg_array=0x7f08d1fc0100, f_reg_array=0x7f08d1fc2180) at x86_64-unknown-linux-gnu/debug/smp/beam_hot.h:310
#8  0x000000000048296c in sched_thread_func (vesdp=0x7f08d0d826c0) at beam/erl_process.c:8349
#9  0x00000000007324b3 in thr_wrapper (vtwd=0x7fffd749d1e0) at pthread/ethread.c:118
#10 0x00007f0912c208ca in start_thread () from /lib/libpthread.so.0
#11 0x00007f091277fb6d in clone () from /lib/libc.so.6
#12 0x0000000000000000 in ?? ()
```